requirements.txt
playwright
jupytext
nbformat
nbconvert
pyyaml
python-dotenv
tzdata
